requirements.txt
numpy==1.23.0
tqdm
pyhocon==0.3.59
opencv-python
trimesh
scikit-learn
scikit-image
plyfile
smplx
pillow-avif-plugin
chumpy
lpips
pandas
einops
wandb
pytorch_wavelets
imageio==2.8.0
imageio-ffmpeg
Pillow==8.4.0
matplotlib
https://github.com/podgorskiy/dnnlib/releases/download/0.0.1/dnnlib-0.0.1-py3-none-any.whl
git+https://github.com/NVlabs/nvdiffrast
git+https://github.com/otaheri/chamfer_distance